Revert "Set error page no index"
